Fixed #46
Code was recognizing that space was pressed, and if space was the last
key pressed it would store it in the key variable.  If you stood still
it would still think you are holding the space down since key == ' '
technically.  To fix it added  a keypress if statement, making it only
detect space if a key was pressed down.